+++ src/main/java/com/wenxin2/marioverse/blocks/PeachAbilityBlock.java
@@ -1,0 +1,27 @@
+package com.wenxin2.marioverse.blocks;
+
+public class PeachAbilityBlock extends AbilityBlock {
+    public PeachAbilityBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public double getNormalJumpBoost() {
+        return 0.3;
+    }
+
+    @Override
+    public double getRunningJumpBoost() {
+        return 0.4;
+    }
+
+    @Override
+    public double getSafeFallDistance() {
+        return 7.0;
+    }
+
+    @Override
+    public double getVerticalMotionMultiplier() {
+        return 0.7;
+    }
+}