+++ src/main/java/com/wenxin2/marioverse/registries/AttributesRegistry.java
@@ -14,17 +14,48 @@ public class AttributesRegistry {
     public static final DeferredHolder<Attribute, Attribute> HEIGHT_SCALE;
     public static final DeferredHolder<Attribute, Attribute> WIDTH_SCALE;
 
-    public static final ResourceLocation AUTO_STEP_HEIGHT = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "auto_step_height");
-    public static final ResourceLocation BLOCK_REACH_DISTANCE = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "block_reach_distance");
-    public static final ResourceLocation COSTUME_ARMOR_KNOCKBACK_RESISTANCE = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "costume_armor_knockback_resistance");
-    public static final ResourceLocation COSTUME_ARMOR_TOUGHNESS = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "costume_armor_toughness");
-    public static final ResourceLocation DAMAGED_SCALE = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "damaged_scale");
-    public static final ResourceLocation ENTITY_REACH_DISTANCE = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "entity_reach_distance");
-    public static final ResourceLocation JUMP_BOOST = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "character_jump_boost");
-    public static final ResourceLocation MAX_HEATH = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "max_health");
-    public static final ResourceLocation MINI_GOOMBA_SLOWDOWN = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "mini_goomba_slowdown");
-    public static final ResourceLocation RUNNING_JUMP_BOOST = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "character_running_jump_boost");
-    public static final ResourceLocation SAFE_FALL_DISTANCE = ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "character_safe_fall_distance");
+    public static final ResourceLocation AUTO_STEP_HEIGHT = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "auto_step_height");
+    public static final ResourceLocation BLOCK_REACH_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "block_reach_distance");
+    public static final ResourceLocation CHARACTER_JUMP_BOOST = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "character_jump_boost");
+    public static final ResourceLocation CHARACTER_RUNNING_JUMP_BOOST = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "character_running_jump_boost");
+    public static final ResourceLocation CHARACTER_SAFE_FALL_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "character_safe_fall_distance");
+    public static final ResourceLocation COSTUME_ARMOR_KNOCKBACK_RESISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "costume_armor_knockback_resistance");
+    public static final ResourceLocation COSTUME_ARMOR_TOUGHNESS = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "costume_armor_toughness");
+    public static final ResourceLocation DAMAGED_SCALE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "damaged_scale");
+    public static final ResourceLocation ENTITY_REACH_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "entity_reach_distance");
+    public static final ResourceLocation MAX_HEATH = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "max_health");
+    public static final ResourceLocation MEGA_BLOCK_REACH_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mega_block_reach_distance");
+    public static final ResourceLocation MEGA_ENTITY_REACH_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mega_entity_reach_distance");
+    public static final ResourceLocation MEGA_JUMP_BOOST = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mega_jump_boost");
+    public static final ResourceLocation MEGA_RUNNING_JUMP_BOOST = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mega_running_jump_boost");
+    public static final ResourceLocation MEGA_SAFE_FALL_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mega_safe_fall_distance");
+    public static final ResourceLocation MINI_BLOCK_REACH_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mini_block_reach_distance");
+    public static final ResourceLocation MINI_ENTITY_REACH_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mini_entity_reach_distance");
+    public static final ResourceLocation MINI_JUMP_BOOST = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mini_jump_boost");
+    public static final ResourceLocation MINI_RUNNING_JUMP_BOOST = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mini_running_jump_boost");
+    public static final ResourceLocation MINI_SAFE_FALL_DISTANCE = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mini_safe_fall_distance");
+    public static final ResourceLocation MINI_GOOMBA_SLOWDOWN = ResourceLocation
+            .fromNamespaceAndPath(Marioverse.MOD_ID, "mini_goomba_slowdown");
 
     static {
         EYE_HEIGHT_SCALE = Marioverse.ATTRIBUTES.register("eye_height_scale",